requirements.txt
# Core ML / Retrieval
rank-bm25==0.2.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy>=1.24.0
torch>=2.0.0

# API Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0

# LLM Generation
openai==1.3.0

# Frontend
streamlit==1.28.0
requests>=2.31.0

# Evaluation & Statistics
scipy>=1.11.0

# Data processing
tqdm>=4.66.0

# Development / Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0  # for FastAPI test client
